admin/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Ezequiel Reale <ezequielreale33@gmail.com>"]
readme = "README.md"
packages = [{include = "web", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12.3"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
flask-session = "^0.8.0"
flask-bcrypt = "^1.0.1"
flask-wtf = "^1.2.1"
wtforms = "^3.1.2"
email-validator = "^2.2.0"
minio = "^7.2.9"
python-ulid = "^2.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
python-dotenv = "^1.0.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
